pyproject.toml
[tool.poetry]
name = "sqlinjectlib"
version = "0.1.0"
description = "A library to simplify SQL injections during CTFs"
authors = ["rikyiso01 <riky.isola@gmail.com>"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Database",
    "Topic :: Utilities",
    "Typing :: Typed",
]
readme="README.md"
repository="https://github.com/rikyiso01/sqlinjectlib"
documentation="https://rikyiso01.github.io/sqlinjectlib"
keywords=["sqlinjection"]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/rikyiso01/sqlinjectlib/issues"

[tool.poetry.dependencies]
python = "^3.10"
typing-extensions = "^4.4.0"

[tool.poetry.group.dev.dependencies]
mysql = "^0.0.3"
pytest = "^7.1.3"
httpx = "^0.23.0"
pytest-asyncio = "^0.20.1"
pytest-lazy-fixture = "^0.6.3"
radon = "^5.1.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
